# openapi.json
{
  "components": {
    "schemas": {
...
      "SimulationBundleResponse": {
        "description": "Full generated skill bundle for the active skill.",
        "properties": {
          "files": {
            "additionalProperties": {
              "type": "string"
            },
            "description": "Bundle filename -> verbatim file contents. Always includes SKILL.md, schema.json, db.json, api.json; scenarios.json is present only when it exists on disk.",
            "title": "Files",
            "type": "object"
          },
          "name": {
            "description": "Name of the active skill/simulation",
            "title": "Name",
            "type": "string"
          },
          "sizes": {
            "additionalProperties": {
              "type": "integer"
            },
            "description": "Bundle filename -> uncompressed content size in bytes.",
            "title": "Sizes",
            "type": "object"
          }
        },
        "required": [
          "name",
          "files",
          "sizes"
        ],
        "title": "SimulationBundleResponse",
        "type": "object"
      },
...
    }
  },
  "info": {
    "contact": {
      "name": "Simulation Harness"
    },
    "description": "Spin up a stateful [MCP](https://modelcontextprotocol.io/) server that simulates any REST API described by an OpenAPI specification \u2014 without hitting real backends.\n\n## Workflow\n\n1. **Create a simulation** \u2014 `POST /api/v1/simulation` with your OpenAPI spec. The harness validates the spec, generates an LLM skill, and starts an MCP server.\n2. **Connect your MCP client** \u2014 point it at the MCP transport endpoint below.\n3. **Call tools** \u2014 each tool maps 1-to-1 with an OpenAPI operation; responses are synthesised by an LLM using the spec's schemas and examples.\n4. **Reset or delete** the simulation when you're done.\n\n## MCP Transports\n\n| Transport | Connect | Messages |\n|-----------|---------|----------|\n| SSE *(default)* | `GET /mcp/sse` | `POST /mcp/messages` |\n| Streamable HTTP | `POST /mcp` | \u2014 |\n\nOnly **one simulation** can be active at a time.",
    "title": "Simulation Harness",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/simulation": {
      "delete": {
        "description": "Cancel in-flight creation or shut down the active simulation. After this call, `GET /simulation` returns 404.",
        "operationId": "delete_simulation_api_v1_simulation_delete",
        "responses": {
          "204": {
            "description": "Simulation deleted (or cancelled mid-creation)"
          },
          "404": {
            "description": "No simulation exists"
          }
        },
        "summary": "Delete simulation",
        "tags": [
          "simulations"
        ]
      },
      "get": {
        "description": "Return the current simulation record including status, progress, and (when ready) session counters. Used by clients to poll until status=ready or failed.",
        "operationId": "get_simulation_api_v1_simulation_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "No simulation has been declared"
          }
        },
        "summary": "Get simulation",
        "tags": [
          "simulations"
        ]
      },
      "post": {
        "description": "Declare a new simulation. Returns immediately with `status=pending` while skill generation and agent initialization run in the background. Poll `GET /api/v1/simulation` until status becomes `ready` or `failed`.\n\nOnly one simulation can be active per process \u2014 a second POST while one exists returns 409.",
        "operationId": "create_simulation_api_v1_simulation_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationResponse"
                }
              }
            },
            "description": "Simulation declared; poll GET /simulation for progress"
          },
          "409": {
            "description": "A simulation already exists \u2014 delete it first"
          },
          "413": {
            "description": "Request body exceeds the 10 MB limit"
          },
          "422": {
            "description": "Invalid OpenAPI specification"
          }
        },
        "summary": "Create simulation (async)",
        "tags": [
          "simulations"
        ]
      }
    },
    "/api/v1/simulation/bundle": {
      "get": {
        "description": "Return every artifact of the active skill's bundle in one response: `SKILL.md`, `schema.json`, `db.json`, `api.json`, and `scenarios.json` when present. Each file is a verbatim string, so writing the values back to a fresh `skills-store/<name>/` reproduces the bundle byte-for-byte and `start` opens a session with no regeneration.\n\n`sizes` reports each file's uncompressed byte length. Send `Accept-Encoding: gzip` to receive a gzip-compressed body (`Content-Encoding: gzip`). Read-only; no session side effects.\n\nNote: a consumer with a size cap (e.g. a 1 MiB Kubernetes ConfigMap) must check the returned sizes itself \u2014 the harness never truncates or rejects on size.",
        "operationId": "get_simulation_bundle_api_v1_simulation_bundle_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationBundleResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "No simulation exists"
          },
          "500": {
            "description": "Skill bundle is incomplete (a required file missing)"
          },
          "503": {
            "description": "Simulation exists but is not ready yet"
          }
        },
        "summary": "Export the full generated skill bundle for the active skill",
        "tags": [
          "simulations"
        ]
      }
    },
    "/api/v1/simulation/database": {
      "get": {
        "description": "Return the contents of the active skill's `db.json` file from disk. This is distinct from `/simulation/state`, which returns the in-memory store snapshot for a thread. Use this to inspect the seed.",
        "operationId": "get_simulation_database_api_v1_simulation_database_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Simulation Database Api V1 Simulation Database Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "No simulation exists"
          },
          "500": {
            "description": "Skill bundle is incomplete (db.json missing)"
          },
          "503": {
            "description": "Simulation exists but is not ready yet"
          }
        },
        "summary": "Get on-disk database (db.json) for the active skill",
        "tags": [
          "simulations"
        ]
      },
      "put": {
        "description": "Atomically overwrite the active skill's `db.json` with the request body. The body is validated against the skill's `schema.json` before the write; validation failures return 422 and the file on disk is unchanged. On success, the running simulation is reset (agent thread and in-memory stores are cleared) so the next `tools/call` reloads from the new seed.\n\nRefuses with 409 while tool calls are in flight.",
        "operationId": "put_simulation_database_api_v1_simulation_database_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "title": "Body",
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Put Simulation Database Api V1 Simulation Database Put",
                  "type": "object"
                }
              }
            },
            "description": "Database replaced; simulation reset"
          },
          "404": {
            "description": "No simulation exists"
          },
          "409": {
            "description": "Tool calls are in flight; retry"
          },
          "422": {
            "description": "Body does not validate against schema.json"
          },
          "503": {
            "description": "Simulation exists but is not ready yet"
          }
        },
        "summary": "Replace on-disk database (db.json) and reset the simulation",
        "tags": [
          "simulations"
        ]
      }
    },
    "/api/v1/simulation/reset": {
      "post": {
        "operationId": "reset_session_api_v1_simulation_reset_post",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": {
                    "type": "string"
                  },
                  "title": "Response Reset Session Api V1 Simulation Reset Post",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "No simulation exists"
          },
          "503": {
            "description": "Simulation exists but is not ready yet"
          }
        },
        "summary": "Reset session",
        "tags": [
          "simulations"
        ]
      }
    },
    "/api/v1/simulation/schema": {
      "get": {
        "description": "Return the contents of the active skill's `schema.json`. Read-only. Clients producing payloads for `PUT /simulation/database` should fetch this first to learn the expected shape.",
        "operationId": "get_simulation_schema_api_v1_simulation_schema_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Simulation Schema Api V1 Simulation Schema Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "No simulation exists"
          },
          "500": {
            "description": "Skill bundle is incomplete (schema.json missing)"
          },
          "503": {
            "description": "Simulation exists but is not ready yet"
          }
        },
        "summary": "Get JSON Schema for the active skill's database",
        "tags": [
          "simulations"
        ]
      }
    },
    "/api/v1/simulation/setup": {
      "post": {
        "description": "Generate the skill artifacts from an OpenAPI spec without starting a session. Returns 202 with status=pending; poll GET /simulation until status=generated or failed.",
        "operationId": "setup_simulation_api_v1_simulation_setup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationResponse"
                }
              }
            },
            "description": "Setup declared; poll GET /simulation for progress"
          },
          "409": {
            "description": "A simulation already exists \u2014 delete it first"
          },
          "413": {
            "description": "Request body exceeds the 10 MB limit"
          },
          "422": {
            "description": "Invalid OpenAPI specification"
          }
        },
        "summary": "Set up a simulation (generate artifacts, no session)",
        "tags": [
          "simulations"
        ]
      }
    },
    "/api/v1/simulation/start": {
      "post": {
        "description": "Open a session from a skill's existing artifacts (no OpenAPI spec required). Returns 202 with status=pending; poll GET /simulation until status=ready or failed.",
        "operationId": "start_simulation_api_v1_simulation_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StartSimulationRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationResponse"
                }
              }
            },
            "description": "Start declared; poll GET /simulation for progress"
          },
          "404": {
            "description": "No complete artifacts exist for the given name"
          },
          "409": {
            "description": "A simulation is already running \u2014 delete it first"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start a simulation from baked artifacts",
        "tags": [
          "simulations"
        ]
      }
    },
    "/api/v1/simulation/state": {
      "get": {
        "operationId": "get_simulation_state_api_v1_simulation_state_get",
        "parameters": [
          {
            "in": "query",
            "name": "thread_id",
            "required": false,
            "schema": {
              "default": "default",
              "title": "Thread Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Get Simulation State Api V1 Simulation State Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "404": {
            "description": "No simulation exists"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          },
          "503": {
            "description": "Simulation exists but is not ready yet"
          }
        },
        "summary": "Get simulation state",
        "tags": [
          "simulations"
        ]
      }
    },
...
  },
...
}